+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.2
+
+- Force cookie to `httponly` is now `true`.
+- Added cookie options `cookie.domain` and `cookie.secure`.
+- Added `cookie.username.key` and `cookie.hash.key` options to have a bit more cryptic cookies.
+
 ## 1.1
 
 - If temp folder is missing it's now added.